Move AnVIL response test to `service` package
